Use DateTimeFormatter for ReshareActionService.parseDateString PR-1099
It'd have been nice to use DateFormatter.parseBest() but the version of
Groovy we're using for Grails 4 doesn't support the :: method reference
so I can't pass eg. `ZonedDateTime::from` as a TemporalQuery. Tried
things like `{ ZonedDateTime.from } as TemporalQuery` also with &. but
no dice so I just worked around it.